:bug: Fix bug releated with the source DB
